+++ packages/card-image/src/Component.screenshots.test.tsx
@@ -1,0 +1,30 @@
+import { setupScreenshotTesting, createSpriteStorybookUrl } from '../../screenshot-utils';
+
+const screenshotTesting = setupScreenshotTesting({
+    it,
+    beforeAll,
+    afterAll,
+    expect,
+});
+
+describe(
+    'CardImage',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'CardImage',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        rounded: [false, true],
+                        width: [350, 500],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);