+++ packages/pure-cell/src/components/content/component.tsx
@@ -1,0 +1,33 @@
+import React, { useContext } from 'react';
+import cn from 'classnames';
+
+import { PureCellContext } from '../../component';
+import { getDataTestId } from '../../../../utils/getDataTestId';
+import { ContentElement } from '../types';
+
+import styles from './index.module.css';
+
+type Props = {
+    /**
+     * Компоненты
+     */
+    children: ContentElement;
+
+    /**
+     * Идентификатор для систем автоматизированного тестирования
+     */
+    dataTestId?: string;
+};
+
+export const Content: React.FC<Props> = ({ children, dataTestId }) => {
+    const { direction } = useContext(PureCellContext);
+
+    return (
+        <section
+            className={cn(styles.component, { [styles.horizontal]: direction === 'horizontal' })}
+            data-test-id={getDataTestId(dataTestId, 'content')}
+        >
+            {children}
+        </section>
+    );
+};