+++ rollup.config.js
@@ -2,6 +2,7 @@ import { ScriptTarget } from 'typescript';
 import path from 'path';
 import multiInput from 'rollup-plugin-multi-input';
 import postcss, { addCssImports, generateClassNameHash } from '@alfalab/rollup-plugin-postcss';
+import wildcardExternal from '@oat-sa/rollup-plugin-wildcard-external';
 import typescript from '@wessberg/rollup-plugin-ts';
 import copy from 'rollup-plugin-copy';
 import json from '@rollup/plugin-json';
@@ -37,6 +38,7 @@ const baseConfig = {
         ...Object.keys(pkg.peerDependencies || {}),
         /(core-js).+/,
     ],
+    plugins: [wildcardExternal(['@alfalab/core-components-*/*'])],
 };
 
 const multiInputPlugin = multiInput();
@@ -76,6 +78,7 @@ const es5 = {
         },
     ],
     plugins: [
+        ...baseConfig.plugins,
         multiInputPlugin,
         typescript({
             tsconfig: resolvedConfig => ({
@@ -109,6 +112,7 @@ const modern = {
         },
     ],
     plugins: [
+        ...baseConfig.plugins,
         multiInputPlugin,
         typescript({
             outDir: 'dist/modern',
@@ -138,6 +142,7 @@ const cssm = {
         },
     ],
     plugins: [
+        ...baseConfig.plugins,
         multiInputPlugin,
         ignoreCss(),
         typescript({
@@ -169,6 +174,7 @@ const esm = {
         },
     ],
     plugins: [
+        ...baseConfig.plugins,
         multiInputPlugin,
         typescript({
             outDir: 'dist/esm',
@@ -192,6 +198,7 @@ const root = {
     input: ['dist/**/*.js'],
     external: baseConfig.external,
     plugins: [
+        ...baseConfig.plugins,
         multiInput({
             relative: 'dist',
         }),