+++ packages/code-input/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-code-input@1.2.3...@alfalab/core-components-code-input@1.3.0) (2022-06-20)
+
+
+### Features
+
+* **code-input:** autocomplete sms in android ([#67](https://github.com/core-ds/core-components/issues/67)) ([b32f734](https://github.com/core-ds/core-components/commit/b32f73403d5ccce9812368e5f186952a67dca57c))
+
+
+
+
+
 # [1.2.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-code-input@1.1.0...@alfalab/core-components-code-input@1.2.0) (2022-03-03)