+++ packages/date-input/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/core-ds/core-components/compare/@alfalab/core-components-date-input@2.2.1...@alfalab/core-components-date-input@2.2.2) (2022-06-23)
+
+**Note:** Version bump only for package @alfalab/core-components-date-input
+
+
+
+
+
+## [2.2.1](https://github.com/core-ds/core-components/compare/@alfalab/core-components-date-input@2.2.0...@alfalab/core-components-date-input@2.2.1) (2022-06-20)
+
+**Note:** Version bump only for package @alfalab/core-components-date-input
+
+
+
+
+
 # [2.2.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-date-input@2.1.5...@alfalab/core-components-date-input@2.2.0) (2022-06-03)