+++ packages/icon-button/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.14](https://github.com/core-ds/core-components/compare/@alfalab/core-components-icon-button@4.1.13...@alfalab/core-components-icon-button@4.1.14) (2022-06-24)
+
+
+### Bug Fixes
+
+* **icon-button:** fix disabled colors ([#104](https://github.com/core-ds/core-components/issues/104)) ([3f03849](https://github.com/core-ds/core-components/commit/3f038495bb63f72cd81ceeedbe55b52119581d57))
+
+
+
+
+
+## [4.1.13](https://github.com/core-ds/core-components/compare/@alfalab/core-components-icon-button@4.1.12...@alfalab/core-components-icon-button@4.1.13) (2022-06-23)
+
+
+### Bug Fixes
+
+* fixed launch storybook in IE ([#52](https://github.com/core-ds/core-components/issues/52)) ([379528b](https://github.com/core-ds/core-components/commit/379528b4ee24183dec38930ea15a31661b994085))
+
+
+
+
+
+## [4.1.12](https://github.com/core-ds/core-components/compare/@alfalab/core-components-icon-button@4.1.11...@alfalab/core-components-icon-button@4.1.12) (2022-06-20)
+
+
+### Bug Fixes
+
+* **icon-button:** fix loading ([#105](https://github.com/core-ds/core-components/issues/105)) ([0b133f0](https://github.com/core-ds/core-components/commit/0b133f042e86702ec2861915f2cdbcdbad9ca905))
+
+
+
+
+
 ## [4.1.11](https://github.com/core-ds/core-components/compare/@alfalab/core-components-icon-button@4.1.10...@alfalab/core-components-icon-button@4.1.11) (2022-06-03)
 
 **Note:** Version bump only for package @alfalab/core-components-icon-button