+++ packages/table/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.11](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.10...@alfalab/core-components-table@1.3.11) (2022-06-24)
+
+**Note:** Version bump only for package @alfalab/core-components-table
+
+
+
+
+
+## [1.3.10](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.9...@alfalab/core-components-table@1.3.10) (2022-06-23)
+
+**Note:** Version bump only for package @alfalab/core-components-table
+
+
+
+
+
+## [1.3.9](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.8...@alfalab/core-components-table@1.3.9) (2022-06-20)
+
+**Note:** Version bump only for package @alfalab/core-components-table
+
+
+
+
+
+## [1.3.8](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.7...@alfalab/core-components-table@1.3.8) (2022-06-20)
+
+**Note:** Version bump only for package @alfalab/core-components-table
+
+
+
+
+
 ## [1.3.7](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.6...@alfalab/core-components-table@1.3.7) (2022-06-16)
 
 **Note:** Version bump only for package @alfalab/core-components-table